works with non-full-width images, collapsed icon with hover effect, works with copy/paste from Eraser
