+++ src/components/EraserLink/eraserlink.css
@@ -1,0 +1,19 @@
+.eraserLinkContainer {
+	position: absolute;
+	top: 10px;
+	right: 10px;
+	height: 29px;
+	width: 43px;
+	transition: width 200ms ease-in-out;
+	overflow: hidden;
+	border: 1px solid #e4e4e4;
+	border-radius: 3px;
+}
+.eraserLinkContainer:hover {
+	width: 149px;
+}
+.eraserLinkContainer img {
+	position: absolute;
+	top: -1px;
+	left: -5px;
+}
